requirements.txt
paho-mqtt==1.6.1
psutil==5.9.5
python-dotenv==1.0.0
discord.py==2.3.2
pyTelegramBotAPI==4.14.0
requests==2.31.0
spotipy==2.23.0